First check-in of the Habit Screen.
Next steps:
- reduce warnings from material3 vs material imports
- implement in the HabitViewModel the snoozeHabit() and doneHabit() functions

Further out, start a settings function where new habits can be added and old ones archived.

Co-pilot tips on why to use object-@Provides vs interface-@Binds:

Dagger/Hilt uses the @Binds annotation to tell it that the implementation of an interface should be provided by a certain class. In your case, @Binds is used to tell Dagger/Hilt that DiaryEntryRepositoryImpl is the implementation of DiaryEntryRepository.  The @Provides annotation, on the other hand, is used to tell Dagger/Hilt how to construct an instance of a class. It's typically used when you need to do some setup or configuration when creating an instance of a class.  In your case, Dagger/Hilt is complaining about duplicate bindings because it's being told two different ways to provide an instance of DiaryEntryRepository: once via @Binds in DataModule.kt, and once via @Provides in RepositoryModule.  The reason Dagger/Hilt prefers the @Binds annotation in DataModule.kt over the @Provides annotation in RepositoryModule is because @Binds is more efficient. When Dagger/Hilt sees a @Binds annotation, it knows it can just use the constructor of the specified class to create an instance. When it sees a @Provides annotation, it has to invoke a method to get an instance.  In general, you should use @Binds when you're binding an implementation to an interface, and @Provides when you need to do some setup or configuration when creating an instance.